added Middleware folder, auth.js, exported middleware to refactor code
